+++ presto-docs/src/main/sphinx/connector.rst
@@ -13,7 +13,7 @@ from different data sources.
     connector/cassandra
     connector/hive
     connector/hive-security
-    connector/inmemory
+    connector/memory
     connector/jmx
     connector/kafka
     connector/kafka-tutorial
