--- presto-in-memory/src/main/java/com/facebook/presto/plugin/inmemory/InMemoryConnector.java
@@ -1,72 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.facebook.presto.plugin.inmemory;
-
-import com.facebook.presto.spi.connector.Connector;
-import com.facebook.presto.spi.connector.ConnectorMetadata;
-import com.facebook.presto.spi.connector.ConnectorPageSinkProvider;
-import com.facebook.presto.spi.connector.ConnectorPageSourceProvider;
-import com.facebook.presto.spi.connector.ConnectorSplitManager;
-import com.facebook.presto.spi.connector.ConnectorTransactionHandle;
-import com.facebook.presto.spi.transaction.IsolationLevel;
-
-public class InMemoryConnector
-        implements Connector
-{
-    private final InMemoryMetadata metadata;
-    private final InMemorySplitManager splitManager;
-    private final InMemoryPageSourceProvider pageSourceProvider;
-    private final InMemoryPageSinkProvider pageSinkProvider;
-
-    public InMemoryConnector(InMemoryMetadata metadata,
-            InMemorySplitManager splitManager,
-            InMemoryPageSourceProvider pageSourceProvider,
-            InMemoryPageSinkProvider pageSinkProvider)
-    {
-        this.metadata = metadata;
-        this.splitManager = splitManager;
-        this.pageSourceProvider = pageSourceProvider;
-        this.pageSinkProvider = pageSinkProvider;
-    }
-
-    @Override
-    public ConnectorTransactionHandle beginTransaction(IsolationLevel isolationLevel, boolean readOnly)
-    {
-        return InMemoryTransactionHandle.INSTANCE;
-    }
-
-    public ConnectorMetadata getMetadata(ConnectorTransactionHandle transactionHandle)
-    {
-        return metadata;
-    }
-
-    @Override
-    public ConnectorSplitManager getSplitManager()
-    {
-        return splitManager;
-    }
-
-    @Override
-    public ConnectorPageSourceProvider getPageSourceProvider()
-    {
-        return pageSourceProvider;
-    }
-
-    @Override
-    public ConnectorPageSinkProvider getPageSinkProvider()
-    {
-        return pageSinkProvider;
-    }
-}
