--- presto-in-memory/src/main/java/com/facebook/presto/plugin/inmemory/InMemoryInsertTableHandle.java
@@ -1,48 +1,0 @@
-/*
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.facebook.presto.plugin.inmemory;
-
-import com.facebook.presto.spi.ConnectorInsertTableHandle;
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import static com.google.common.base.MoreObjects.toStringHelper;
-import static java.util.Objects.requireNonNull;
-
-public final class InMemoryInsertTableHandle
-        implements ConnectorInsertTableHandle
-{
-    private final InMemoryTableHandle table;
-
-    @JsonCreator
-    public InMemoryInsertTableHandle(@JsonProperty("table") InMemoryTableHandle table)
-    {
-        this.table = requireNonNull(table, "table is null");
-    }
-
-    @JsonProperty
-    public InMemoryTableHandle getTable()
-    {
-        return table;
-    }
-
-    @Override
-    public String toString()
-    {
-        return toStringHelper(this)
-                .add("table", table)
-                .toString();
-    }
-}
